backend/requirements.txt
flask==3.0.3
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pillow==10.4.0
openai==1.30.0
httpx==0.28.1
python-dotenv==1.0.0
flask-cors==4.0.0
werkzeug==3.0.1
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0
pytesseract==0.3.13
pymupdf==1.24.10
